requirements.txt
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.3
coverage==6.3.1
cryptography==36.0.1
Flask==2.0.3
Flask-Cors==3.0.10
flask-ssl==0.0.0.1rc1
Flask-SSLify==0.1.5
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mypy==0.931
mypy-extensions==0.4.3
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.3.0
python-dotenv==0.19.2
requests==2.27.1
six==1.16.0
tomli==2.0.1
typing-extensions==4.1.1
urllib3==1.26.8
Werkzeug==2.0.3
